# perf: improve retrieving `versionInfo` on Turbo HMR