@ijjk 2024-06-29
## Failing test suites <!--25462409372-->
Commit: 93552ff4d8423e48bb6be366873e61897b40fd45

`TURBOPACK=1 pnpm test test/integration/app-document-remove-hmr/test/index.test.js` (turbopack)
- _app removal HMR > should HMR when _app is removed

<details>
<summary>Expand output</summary>

  ● _app removal HMR › should HMR when _app is removed

    TIMED OUT: success

    <head><meta charset="utf-8" data-next-head=""><meta name="viewport" content="width=device-width" data-next-head=""><noscript data-n-css=""></noscript><script src="/_next/static/chunks/%5Broot%20of%20the%20server%5D__fea3e8._.js" defer=""></script><script src="/_next/static/chunks/9752e_react_a91040._.js" defer=""></script><script src="/_next/static/chunks/5363e_react-dom_cjs_react-dom-client_development_26dbd4.js" defer=""></script><script src="/_next/static/chunks/5363e_react-dom_cjs_react-dom_development_2a980d.js" defer=""></script><script src="/_next/static/chunks/5363e_react-dom_609578._.js" defer=""></script><script src="/_next/static/chunks/08b5e__pnpm_798ac5._.js" defer=""></script><script src="/_next/static/chunks/%5Bproject%5D_packages_next_dist_pages_2ad8dc._.js" defer=""></script><script src="/_next/static/chunks/%5Bnext%5D_entry_page-loader_ts_b905c6._.js" defer=""></script><script src="/_next/static/chunks/%5Bnext%5D_entry_page-loader_ts_418caf._.js" defer=""></script><script src="/_next/static/chunks/%5Broot%20of%20the%20server%5D__495e86._.js" defer=""></script><script src="/_next/static/chunks/9752e_react_a51a5e._.js" defer=""></script><script src="/_next/static/chunks/%5Bproject%5D_packages_next_dist_pages_3d95ef._.js" defer=""></script><script src="/_next/static/chunks/%5Bnext%5D_entry_page-loader_ts_19f94b._.js" defer=""></script><script src="/_next/static/chunks/%5Bnext%5D_entry_page-loader_ts_f86330._.js" defer=""></script><script src="/_next/static/development/_ssgManifest.js" defer=""></script><script src="/_next/static/development/_buildManifest.js" defer=""></script><noscript id="__next_css__DO_NOT_USE__"></noscript></head><body><p>custom _document</p><div id="__next"><p>index page</p></div><script id="__NEXT_DATA__" type="application/json">{"props":{"pageProps":{}},"page":"/","query":{},"buildId":"development","nextExport":true,"autoExport":true,"isFallback":false,"scriptLoader":[]}</script><div id="__next-build-watcher" style="position: fixed; bottom: 10px; right: 20px; width: 0px; height: 0px; z-index: 99999;"></div><next-route-announcer><p aria-live="assertive" id="__next-route-announcer__" role="alert" style="border: 0px; clip: rect(0px, 0px, 0px, 0px); height: 1px; margin: -1px; overflow: hidden; padding: 0px; position: absolute; top: 0px; width: 1px; white-space: nowrap; overflow-wrap: normal;"></p></next-route-announcer></body>

    undefined

      699 |
      700 |   if (hardError) {
    > 701 |     throw new Error('TIMED OUT: ' + regex + '\n\n' + content + '\n\n' + lastErr)
          |           ^
      702 |   }
      703 |   return false
      704 | }

      at check (lib/next-test-utils.ts:701:11)
      at Object.<anonymous> (integration/app-document-remove-hmr/test/index.test.js:49:7)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).